requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
appscript==1.2.2
asyncpg==0.30.0
attrs==23.1.0
autobahn==19.11.2
Automat==22.10.0
bcrypt==4.0.1
blinker==1.9.0
boltons==23.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.8
constantly==15.1.0
cryptography==41.0.1
docopt==0.6.2
environs==9.5.0
fastapi==0.115.12
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.1.1
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
greenlet==3.2.3
gspread==5.10.0
gspread-dataframe==3.3.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kiteconnect==4.2.0
lxml==4.9.2
MarkupSafe==3.0.2
marshmallow==3.19.0
msgpack==1.0.5
numpy==1.25.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pendulum==2.1.2
pluggy==1.5.0
protobuf==4.23.3
psutil==5.9.5
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.11.6
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==23.2.0
pyparsing==3.1.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
service-identity==23.1.0
six==1.16.0
sniffio==1.3.1
sourcedefender==10.0.13
SQLAlchemy==2.0.20
starlette==0.46.2
TgCrypto==1.2.5
Twisted==22.10.0
txaio==23.1.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
xlwings==0.30.10
zope.interface==6.0
